fsm output
