+++ design/pycde_analysis/FSMUser/hw/FSMUser.sv
@@ -1,0 +1,26 @@
+// Generated by CIRCT unknown git version
+`include "fsm_enum_typedefs.sv"
+module FSMUser
+  #(parameter __INST_HIER = "INSTANTIATE_WITH_INSTANCE_PATH") (	// /home/kerrwang/repos/testing1/design/pycde_analysis/./mmain.py:105
+  input  a,
+         b,
+         c,
+         clk,
+         rst,
+  output is_a,
+         is_b,
+         is_c
+);
+
+  F0 F0 (	// /home/kerrwang/repos/testing1/design/pycde_analysis/./mmain.py:116
+    .in0  (a),
+    .in1  (b),
+    .in2  (c),
+    .clk  (clk),
+    .rst  (rst),
+    .out0 (/* unused */),
+    .out1 (is_a),
+    .out2 (is_b),
+    .out3 (is_c)
+  );
+endmodule
