+++ design/pycde_analysis/FSMUser/hw/filelist.f
@@ -1,0 +1,5 @@
+fsm_enum_typedefs.sv
+FSMUser.sv
+F0.sv
+services.json
+FSMUser.tcl
