+++ design/pycde_analysis/FSMUser/hw/fsm_enum_typedefs.sv
@@ -1,0 +1,6 @@
+// Generated by CIRCT unknown git version
+`ifndef _TYPESCOPE_fsm_enum_typedecls
+`define _TYPESCOPE_fsm_enum_typedecls
+  typedef enum bit [1:0] {F0_state_t_idle, F0_state_t_A, F0_state_t_B, F0_state_t_C}
+    F0_state_t;	// /home/kerrwang/repos/testing1/design/pycde_analysis/./mmain.py:78
+`endif // _TYPESCOPE_fsm_enum_typedecls
