+++ design/pycde_analysis/FSMUser/hw/services.json
@@ -1,0 +1,10 @@
+{
+  "declarations": [],
+  "top_levels": [
+    {
+      "module": "@FSMUser",
+      "services": []
+    }
+  ],
+  "modules": []
+}
