+++ design/pycde_analysis/exsys/hw/OrInts.sv
@@ -1,0 +1,9 @@
+// Generated by CIRCT unknown git version
+module OrInts
+  #(parameter __INST_HIER = "INSTANTIATE_WITH_INSTANCE_PATH") (	// /home/kerrwang/repos/testing1/design/pycde_analysis/./mmain.py:5
+  input  [31:0] a,
+                b,
+  output [31:0] c);
+
+  assign c = a | b;	// /home/kerrwang/repos/testing1/design/pycde_analysis/./mmain.py:11, :12
+endmodule
